added grid lay to map app
