+++ frontend/regional-mma2/src/App.css
@@ -1,9 +1,18 @@
 .App {
   text-align: center;
+  display:grid;
+  grid-template-columns: 2fr 2fr;
+  
 }
 
 .Content{
   margin: 0 auto;
   text-align: center;
   max-width:960px;
+
+}
+
+.Prospects_Section{
+  padding:0;
+  margin:0;
 }
