+++ frontend/regional-mma2/src/Modal.css
@@ -10,4 +10,5 @@
     font-family:sans-serif;
     color:white;
     letter-spacing:.2em;
+    border-radius:5px;
 }
