+++ frontend/regional-mma2/src/Modal.js
@@ -4,7 +4,6 @@ import './Modal.css';
 
 const Modal =(props)=>(    
     <div className="Modal__Wrapper">
-        <Loader type="ThreeDots" color="#somecolor" height={80} width={80} />
     </div>
 )
